Add Regolith support
Moved core logic to a new package directory (minecraft_lang_tool/) and split CLI functionality into cli.py. Removed the monolithic minecraft_lang_tool.py file and reorganized code for better maintainability and modularity. Added filter.json for configuration.
